requirements.txt
flask>=3.1.0
pillow>=11.1.0
torch>=2.6.0
torchvision>=0.21.0
numpy>=2.2.4
pandas>=2.2.3
scikit-learn>=1.6.1
matplotlib>=3.10.1
tqdm>=4.67.1
albumentations>=2.0.5
opencv-python-headless>=4.9.0
requests>=2.31.0